There is a bug with not displaying the player

+++ src/com/game/main/Block.java
@@ -1,6 +1,5 @@
 package com.game.main;
 
-import sun.plugin2.util.ColorUtil;
 
 import java.awt.*;
 
@@ -18,12 +17,12 @@ public class Block extends GameObject {
     public void render(Graphics g) {
 
         g.setColor(Color.DARK_GRAY);
-        g.fillRect(x, y, 32, 32);
-        g.drawRect(x, y, 32, 32);
+        g.fillRect((int)x, (int)y, 32, 32);
+        g.drawRect((int)x, (int)y, 32, 32);
     }
 
 
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 32 ,32);
+        return new Rectangle((int)x, (int)y, 32 ,32);
     }
 }
